UIManager 생성 마우스 클릭에 대한 UI 반응추가 하지만, 지금같은 경우 항상 부모가 반응

+++ WinApi2D/CScene.cpp
@@ -90,7 +90,7 @@ void CScene::render_tile(HDC hDC)
     {
         for (int iCurCol = iLTCol; iCurCol <= (iLTCol + iClientWidth); ++iCurCol)
         {
-            if (iCurCol < 0 || m_iTileX <= iCurCol || iCurRow < 0 || m_iTileY <= iCurRow)
+            if (iCurCol < 0 || m_iTileX <= (UINT)iCurCol || iCurRow < 0 || m_iTileY <= (UINT)iCurRow)
             {
                 continue;
             }
